Changed runner to be raw output, added RunnerOutput
