+++ buildSrc/src/main/groovy/common-conventions.gradle
@@ -29,7 +29,7 @@ publishing {
         maven(MavenPublication) {
             groupId = 'edu.austral.ingsis'
             artifactId = "${project.name}"
-            version = '1.0.9'
+            version = '1.0.10'
 
             from(components.java)
         }
